flasher: manual update to latest ak3 osm0sis/AnyKernel3@d3d853ac825561c74a0a0e58b63b5b6c90420e90

+++ flasher/tools/ak3-core.sh
@@ -11,25 +11,18 @@ patch=$home/patch;
 ramdisk=$home/ramdisk;
 split_img=$home/split_img;
 
-$BOOTMODE && cpio=$bin/cpio || cpio=cpio
-
 ### output/testing functions:
 # ui_print "<text>" [...]
 ui_print() {
   until [ ! "$1" ]; do
-    if $BOOTMODE; then
-      echo "$1";
-    else
-      echo -e "ui_print $1\nui_print" > /proc/self/fd/$OUTFD;
-    fi;
-
+    echo -e "ui_print $1\nui_print" > /proc/self/fd/$OUTFD;
     shift;
   done;
 }
 
 # abort ["<text>" [...]]
 abort() {
-  ui_print " " "!! $*; aborting." " ";
+  ui_print " " "$@";
   exit 1;
 }
 
@@ -44,13 +37,37 @@ file_getprop() {
 }
 ###
 
+### file/directory attributes functions:
+# set_perm <owner> <group> <mode> <file> [<file2> ...]
+set_perm() {
+  local uid gid mod;
+  uid=$1; gid=$2; mod=$3;
+  shift 3;
+  chown $uid:$gid "$@" || chown $uid.$gid "$@";
+  chmod $mod "$@";
+}
+
+# set_perm_recursive <owner> <group> <dir_mode> <file_mode> <dir> [<dir2> ...]
+set_perm_recursive() {
+  local uid gid dmod fmod;
+  uid=$1; gid=$2; dmod=$3; fmod=$4;
+  shift 4;
+  while [ "$1" ]; do
+    chown -R $uid:$gid "$1" || chown -R $uid.$gid "$1";
+    find "$1" -type d -exec chmod $dmod {} +;
+    find "$1" -type f -exec chmod $fmod {} +;
+    shift;
+  done;
+}
+###
+
 ### dump_boot functions:
 # split_boot (dump and split image only)
 split_boot() {
   local dumpfail;
 
   if [ ! -e "$(echo $block | cut -d\  -f1)" ]; then
-    abort "Invalid partition";
+    abort "Invalid partition. Aborting...";
   fi;
   if [ "$(echo $block | grep ' ')" ]; then
     block=$(echo $block | cut -d\  -f1);
@@ -116,7 +133,7 @@ split_boot() {
   fi;
 
   if [ $? != 0 -o "$dumpfail" ]; then
-    abort "Dumping/splitting image failed";
+    abort "Dumping/splitting image failed. Aborting...";
   fi;
   cd $home;
 }
@@ -137,13 +154,13 @@ unpack_ramdisk() {
   if [ -f ramdisk.cpio ]; then
     comp=$($bin/magiskboot decompress ramdisk.cpio 2>&1 | grep -v 'raw' | sed -n 's;.*\[\(.*\)\];\1;p');
   else
-    abort "No ramdisk found to unpack"
+    abort "No ramdisk found to unpack. Aborting...";
   fi;
   if [ "$comp" ]; then
     mv -f ramdisk.cpio ramdisk.cpio.$comp;
     $bin/magiskboot decompress ramdisk.cpio.$comp ramdisk.cpio;
     if [ $? != 0 ]; then
-      echo "Attempting with busybox $comp..." >&2;
+      echo "Attempting ramdisk unpack with busybox $comp..." >&2;
       $comp -dc ramdisk.cpio.$comp > ramdisk.cpio;
     fi;
   fi;
@@ -153,9 +170,9 @@ unpack_ramdisk() {
   chmod 755 $ramdisk;
 
   cd $ramdisk;
-  EXTRACT_UNSAFE_SYMLINKS=1 $cpio -d -F $split_img/ramdisk.cpio -i;
+  EXTRACT_UNSAFE_SYMLINKS=1 cpio -d -F $split_img/ramdisk.cpio -i;
   if [ $? != 0 -o ! "$(ls)" ]; then
-    abort "Unpacking ramdisk failed"
+    abort "Unpacking ramdisk failed. Aborting...";
   fi;
   if [ -d "$home/rdtmp" ]; then
     cp -af $home/rdtmp/* .;
@@ -163,10 +180,8 @@ unpack_ramdisk() {
 }
 ### dump_boot (dump and split image, then extract ramdisk)
 dump_boot() {
-  ui_print "  • Unpacking boot image";
   split_boot;
-  #ui_print "  • Unpacking ramdisk";
-  #unpack_ramdisk;
+  unpack_ramdisk;
 }
 ###
 
@@ -190,22 +205,25 @@ repack_ramdisk() {
     $bin/mkbootfs $ramdisk > ramdisk-new.cpio;
   else
     cd $ramdisk;
-    find . | $cpio -H newc -o > $home/ramdisk-new.cpio;
+    find . | cpio -H newc -o > $home/ramdisk-new.cpio;
   fi;
   test $? != 0 && packfail=1;
 
   cd $home;
+  $bin/magiskboot cpio ramdisk-new.cpio test;
+  magisk_patched=$?;
+  test $((magisk_patched & 3)) -eq 1 && $bin/magiskboot cpio ramdisk-new.cpio "extract .backup/.magisk $split_img/.magisk";
   if [ "$comp" ]; then
     $bin/magiskboot compress=$comp ramdisk-new.cpio;
     if [ $? != 0 ]; then
-      echo "Attempting with busybox $comp..." >&2;
+      echo "Attempting ramdisk repack with busybox $comp..." >&2;
       $comp -9c ramdisk-new.cpio > ramdisk-new.cpio.$comp;
       test $? != 0 && packfail=1;
       rm -f ramdisk-new.cpio;
     fi;
   fi;
   if [ "$packfail" ]; then
-    abort "Repacking ramdisk failed"
+    abort "Repacking ramdisk failed. Aborting...";
   fi;
 
   if [ -f "$bin/mkmtkhdr" -a -f "$split_img/boot.img-base" ]; then
@@ -218,9 +236,7 @@ repack_ramdisk() {
 
 # flash_boot (build, sign and write image only)
 flash_boot() {
-  local varlist kernel ramdisk cmdline part0 part1 nocompflag signfail pk8 cert avbtype;
-
-  ui_print "  • Repacking boot image";
+  local varlist i kernel ramdisk fdt cmdline comp part0 part1 nocompflag signfail pk8 cert avbtype;
 
   cd $split_img;
   if [ -f "$bin/mkimage" ]; then
@@ -289,22 +305,55 @@ flash_boot() {
     test "$dt" && dt="--dt $dt";
     $bin/mkbootimg --kernel $kernel --ramdisk $ramdisk --cmdline "$cmdline" --base $home --pagesize $pagesize --kernel_offset $kerneloff --ramdisk_offset $ramdiskoff --tags_offset "$tagsoff" $dt --output $home/boot-new.img;
   else
-    test "$kernel" && cp -f $kernel kernel;
+    if [ ! "$magisk_patched" ]; then
+      $bin/magiskboot cpio ramdisk.cpio test;
+      magisk_patched=$?;
+      test $((magisk_patched & 3)) -eq 1 && $bin/magiskboot cpio ramdisk.cpio "extract .backup/.magisk .magisk";
+    fi;
+    if [ "$kernel" ]; then
+      cp -f $kernel kernel;
+      if [ $((magisk_patched & 3)) -eq 1 ]; then
+        ui_print " " "Magisk detected! Patching kernel so reflashing Magisk is not necessary...";
+        comp=$($bin/magiskboot decompress kernel 2>&1 | grep -v 'raw' | sed -n 's;.*\[\(.*\)\];\1;p');
+        ($bin/magiskboot split $kernel || $bin/magiskboot decompress $kernel kernel) 2>/dev/null;
+        if [ $? != 0 -a "$comp" ]; then
+          echo "Attempting kernel unpack with busybox $comp..." >&2;
+          $comp -dc $kernel > kernel;
+        fi;
+        $bin/magiskboot hexpatch kernel 736B69705F696E697472616D667300 77616E745F696E697472616D667300;
+        if [ "$comp" ]; then
+          $bin/magiskboot compress=$comp kernel kernel.$comp;
+          if [ $? != 0 ]; then
+            echo "Attempting kernel repack with busybox $comp..." >&2;
+            $comp -9c kernel > kernel.$comp;
+          fi;
+          mv -f kernel.$comp kernel;
+        fi;
+      else
+        case $kernel in
+          *-dtb) rm -f kernel_dtb;;
+        esac;
+      fi;
+    fi;
     test "$ramdisk" && cp -f $ramdisk ramdisk.cpio;
-    case $kernel in
-      *-dtb) rm -f kernel_dtb;;
-    esac;
     test "$dt" -a -f extra && cp -f $dt extra;
     for i in dtb recovery_dtbo; do
       test "$(eval echo \$$i)" -a -f $i && cp -f $(eval echo \$$i) $i;
     done;
+    if [ $((magisk_patched & 3)) -eq 1 ]; then
+      export $(cat .magisk);
+      test $((magisk_patched & 8)) -ne 0 && export TWOSTAGEINIT=true;
+      for fdt in dtb extra kernel_dtb recovery_dtbo; do
+        test -f $fdt && $bin/magiskboot dtb $fdt patch;
+      done;
+    fi;
     case $ramdisk_compression in
       none|cpio) nocompflag="-n";;
     esac;
     $bin/magiskboot repack $nocompflag $bootimg $home/boot-new.img;
   fi;
   if [ $? != 0 ]; then
-    abort "Repacking image failed"
+    abort "Repacking image failed. Aborting...";
   fi;
 
   cd $home;
@@ -321,22 +370,20 @@ flash_boot() {
       *recovery*|*SOS*) avbtype=recovery;;
       *) avbtype=boot;;
     esac;
-    if [ "$(/system/bin/dalvikvm -Xbootclasspath:/system/framework/core-oj.jar:/system/framework/core-libart.jar:/system/framework/conscrypt.jar:/system/framework/bouncycastle.jar -Xnodex2oat -Xnoimage-dex2oat -cp $bin/BootSignature_Android.jar com.android.verity.BootSignature -verify boot.img 2>&1 | grep VALID)" ]; then
-      /system/bin/dalvikvm -Xbootclasspath:/system/framework/core-oj.jar:/system/framework/core-libart.jar:/system/framework/conscrypt.jar:/system/framework/bouncycastle.jar -Xnodex2oat -Xnoimage-dex2oat -cp $bin/BootSignature_Android.jar com.android.verity.BootSignature /$avbtype boot-new.img $pk8 $cert boot-new-signed.img;
+    if [ "$(/system/bin/dalvikvm -Xnoimage-dex2oat -cp $bin/BootSignature_Android.jar com.android.verity.BootSignature -verify boot.img 2>&1 | grep VALID)" ]; then
+      /system/bin/dalvikvm -Xnoimage-dex2oat -cp $bin/BootSignature_Android.jar com.android.verity.BootSignature /$avbtype boot-new.img $pk8 $cert boot-new-signed.img;
     fi;
   fi;
   if [ $? != 0 -o "$signfail" ]; then
-    abort "Signing image failed"
+    abort "Signing image failed. Aborting...";
   fi;
   mv -f boot-new-signed.img boot-new.img 2>/dev/null;
 
   if [ ! -f boot-new.img ]; then
-    abort "No repacked image found to flash"
+    abort "No repacked image found to flash. Aborting...";
   elif [ "$(wc -c < boot-new.img)" -gt "$(wc -c < boot.img)" ]; then
-    abort "New image larger than boot partition"
+    abort "New image larger than boot partition. Aborting...";
   fi;
-
-  ui_print "  • Flashing new boot image";
   if [ -f "$bin/flash_erase" -a -f "$bin/nandwrite" ]; then
     $bin/flash_erase $block 0 0;
     $bin/nandwrite -p $block boot-new.img;
@@ -347,7 +394,7 @@ flash_boot() {
     cat boot-new.img /dev/zero > $block 2>/dev/null || true;
   fi;
   if [ $? != 0 ]; then
-    abort "Flashing image failed"
+    abort "Flashing image failed. Aborting...";
   fi;
 }
 
@@ -366,7 +413,7 @@ flash_dtbo() {
   if [ "$dtbo" ]; then
     dtboblock=/dev/block/bootdevice/by-name/dtbo$slot;
     if [ ! -e "$dtboblock" ]; then
-      abort "dtbo partition could not be found"
+      abort "dtbo partition could not be found. Aborting...";
     fi;
     if [ -f "$bin/flash_erase" -a -f "$bin/nandwrite" ]; then
       $bin/flash_erase $dtboblock 0 0;
@@ -378,14 +425,13 @@ flash_dtbo() {
       cat $dtbo /dev/zero > $dtboblock 2>/dev/null || true;
     fi;
     if [ $? != 0 ]; then
-      abort "Flashing dtbo failed"
+      abort "Flashing dtbo failed. Aborting...";
     fi;
   fi;
 }
 ### write_boot (repack ramdisk then build, sign and write image and dtbo)
 write_boot() {
-  #ui_print "  • Repacking ramdisk";
-  #repack_ramdisk;
+  repack_ramdisk;
   flash_boot;
   flash_dtbo;
 }
@@ -596,7 +642,8 @@ reset_ak() {
       test -e $i && cp -af $i $current;
     done;
   fi;
-  rm -rf $bootimg $ramdisk $split_img $home/*-new* $home/*-files/current;
+  test -d $split_img && rm -rf $ramdisk;
+  rm -rf $bootimg $split_img $home/*-new* $home/*-files/current;
 
   if [ "$1" == "keep" ]; then
     test -d $home/rdtmp && mv -f $home/rdtmp $ramdisk;
@@ -642,7 +689,7 @@ setup_ak() {
         esac;
       fi;
       if [ ! "$slot" -a "$is_slot_device" == 1 ]; then
-        abort "Unable to determine active boot slot"
+        abort "Unable to determine active boot slot. Aborting...";
       fi;
     ;;
   esac;
@@ -665,9 +712,11 @@ setup_ak() {
             if [ "$mtdpart" == "$part" ]; then
               mtdname=$(echo $mtdmount | cut -d: -f1);
             else
-              abort "Unable to determine mtd $block partition"
+              abort "Unable to determine mtd $block partition. Aborting...";
             fi;
-            target=/dev/mtd/$mtdname;
+            if [ -e /dev/mtd/$mtdname ]; then
+              target=/dev/mtd/$mtdname;
+            fi;
           elif [ -e /dev/block/by-name/$part ]; then
             target=/dev/block/by-name/$part;
           elif [ -e /dev/block/bootdevice/by-name/$part ]; then
@@ -679,13 +728,13 @@ setup_ak() {
           elif [ -e /dev/$part ]; then
             target=/dev/$part;
           fi;
-          test -e "$target" && break 2;
+          test "$target" && break 2;
         done;
       done;
       if [ "$target" ]; then
         block=$(ls $target 2>/dev/null);
       else
-        abort "Unable to determine $block partition"
+        abort "Unable to determine $block partition. Aborting...";
       fi;
     ;;
     *)
